完善对比

+++ src/main/java/com/wonderful/bean/entity/AttacherCardSaleComparisonDetails.java
@@ -45,7 +45,7 @@ public class AttacherCardSaleComparisonDetails {
 
     private BigDecimal payAmount;
 
-    private Integer discount;
+    private BigDecimal discount;
 
     private BigDecimal realPayAmount;
